Choose the remote more wisely
If a branch has a remote branch to track, we display its URL, which is
typically that of a fork. If there is no remote branch, we fall back
to "origin/master", which at least it is guaranteed to work.

As for the utility of this, if I'm working on a branch that I haven't
pushed yet, the old version gives an invalid URL. For instance, for my
currently not-pushed branch:

    $ LANG=C git rev-parse --abbrev-ref HEAD@{push}
    fatal: no upstream configured for branch 'iadaiada'
    # This URL does not exist!
    $ git-url README.md
    https://github.com/Piojo/qontract-reconcile/blob/iadaiada/README.md
    # But this one does
    $ bash $this_repo/git-url README.md
    https://github.com/Piojo/qontract-reconcile/blob/master/README.md
    # And now this works as expected
    $ bash ~/paso/git-url/git-url README.md upstream
    https://github.com/app-sre/qontract-reconcile/blob/master/README.md